myrequirement.txt
backports.entry-points-selectable==1.1.0
click==8.0.1
colorama==0.4.4
distlib==0.3.2
dnspython==2.1.0
email-validator==1.1.3
filelock==3.0.12
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysql-connector==2.2.9
mysql-connector-python==8.0.26
parameters-validation==1.2.0
platformdirs==2.3.0
protobuf==3.17.3
python-dotenv==0.19.0
regex-generator-lib==1.0.0
six==1.16.0
SQLAlchemy==1.4.23
virtualenv==20.7.2
Werkzeug==2.0.1
WTForms==2.3.3
xmltodict==0.12.0
